+++ modeling/parser/expressions/Expression.java
@@ -1,0 +1,3 @@
+package modeling.parser.expressions;
+
+public interface Expression { }