+++ modeling/parser/expressions/LiteralExpression.java
@@ -1,0 +1,9 @@
+package modeling.parser.expressions;
+
+public class LiteralExpression implements Expression {
+    private final Object value;
+
+    public LiteralExpression(Object value) {
+        this.value = value;
+    }
+}