+++ modeling/parser/statements/IfStatement.java
@@ -1,0 +1,15 @@
+package modeling.parser.statements;
+
+import modeling.parser.expressions.Expression;
+
+public class IfStatement implements Statement {
+    private final Expression condition;
+    private final Statement thenBranch;
+    private final Statement elseBranch;
+
+    public IfStatement(Expression condition, Statement thenBranch, Statement elseBranch) {
+        this.condition = condition;
+        this.thenBranch = thenBranch;
+        this.elseBranch = elseBranch;
+    }
+}