+++ modeling/parser/statements/Statement.java
@@ -1,0 +1,3 @@
+package modeling.parser.statements;
+
+public interface Statement { }